refactor(ucd-store-v2): consolidate snapshot management into lockfile module
Merged snapshot management functions into the lockfile module, enhancing the organization of related functionalities. This update includes the addition of methods for reading, writing, and validating snapshots, as well as utility functions for hash computation and path retrieval. The previous snapshot.ts file has been removed to streamline the codebase and improve maintainability.
